requirements.txt
# Core dependencies
langroid>=0.58.0
transformers>=4.53.0
neo4j>=5.28.0
autogen>=0.9.6
faiss-cpu>=1.11.0
sentence-transformers>=5.0.0
python-dotenv>=1.0.0

# # Optional dependencies for Jupyter Notebook support
# notebook>=7.7.4
# ipykernel>=6.30.0